requirements.txt
Django>=5.0
djangorestframework>=3.14
python-dotenv>=1.0
drf-spectacular>=0.27